+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
-- Add `OpenapiParameters::Query#unknown_values` to find unknown query parameters 
+- remove `OpenapiParameters::Parameter#definition` to clean
+- Add `OpenapiParameters::Query#unknown_values` to find unknown query parameters
 
 ## [0.9.0] - 2025-11-07